+++ cmd/local/ddcio/ddcio_test.go
@@ -153,7 +153,7 @@ func TestEnsureClose(t *testing.T) {
 	expectedFile := "my_long_file_name.txt"
 
 	// so the simplelogger output will be captured
-	simplelog.InitLogger(2)
+	simplelog.InitLogger()
 	ddcio.EnsureClose(expectedFile, failedClose)
 
 	raw, err := os.ReadFile(simplelog.GetLogLoc())