+++ pkg/archive/archive.go
@@ -34,9 +34,10 @@ func TarGzDir(srcDir, dest string) error {
 func TarDDC(srcDir, dest, baseDDC string) error {
 	summaryJSON := filepath.Join(srcDir, "summary.json")
 	ddcFolder := filepath.Join(srcDir, baseDDC)
+	simplelog.Debug("copying log to archive for diagnostics")
 	err := simplelog.CopyLog(filepath.Join(baseDDC, "ddc.log"))
 	if err != nil {
-		simplelog.Warningf("unable to copy ddc.log: \n%v", err)
+		fmt.Printf("unable to copy ddc.log: \n%v", err)
 	}
 
 	return TarGzDirFiltered(srcDir, dest, func(name string) bool {